Library Management System
Library Management
Ordering a book online making sure that you have entered valid details the of the customer. Using python making sure that once you entered your details it will tell you if the book is available and if you didn't put the correct order in it will show the books is already taken through a message box.